# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  psychologist: 'psychologist',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface LinkRequest {
  code: string;
}

export interface LinkResult {
  psychologistName: string;
  psychologistId: string;
}

export interface LinkCode {
  code: string;
  expiresAt: string;
}

export interface MoodEntry {
  id: string;
  userId: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  moodScore: number;
  emotions: string[];
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface MoodEntryOrNull {
  entry: MoodEntry | null;
}

export interface MoodEntriesResponse {
  entries: MoodEntry[];
  total: number;
}

export interface CreateMoodEntryRequest {
  /**
     * @minimum 1
     * @maximum 10
     */
  moodScore: number;
  emotions: string[];
  /** @nullable */
  note?: string | null;
}

export interface AffirmationResponse {
  text: string;
  date: string;
}

export interface Task {
  id: string;
  userId: string;
  psychologistId: string;
  title: string;
  description: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface TaskInput {
  title: string;
  description?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface TasksResponse {
  tasks: Task[];
}

export type ClientSummaryTrend = typeof ClientSummaryTrend[keyof typeof ClientSummaryTrend];


export const ClientSummaryTrend = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export type ClientSummaryRisk = typeof ClientSummaryRisk[keyof typeof ClientSummaryRisk];


export const ClientSummaryRisk = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ClientSummary {
  id: string;
  name: string;
  email: string;
  lastCheckin: string;
  avgMood: number;
  trend: ClientSummaryTrend;
  risk: ClientSummaryRisk;
  streak: number;
  emotions: string[];
}

export interface ClientSummariesResponse {
  clients: ClientSummary[];
}

export type RiskAlertType = typeof RiskAlertType[keyof typeof RiskAlertType];


export const RiskAlertType = {
  mood_drop: 'mood_drop',
  no_checkin: 'no_checkin',
  distress_emotion: 'distress_emotion',
  task_overdue: 'task_overdue',
} as const;

export type RiskAlertSeverity = typeof RiskAlertSeverity[keyof typeof RiskAlertSeverity];


export const RiskAlertSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface RiskAlert {
  id: string;
  psychologistId: string;
  clientId: string;
  clientName: string;
  type: RiskAlertType;
  message: string;
  severity: RiskAlertSeverity;
  reviewed: boolean;
  createdAt: string;
}

export interface AlertsResponse {
  alerts: RiskAlert[];
}

export interface JournalEntry {
  id: string;
  userId: string;
  title: string;
  body: string;
  /** @nullable */
  moodTag?: string | null;
  isShared: boolean;
  createdAt: string;
}

export interface CreateJournalEntryRequest {
  title?: string;
  body: string;
  /** @nullable */
  moodTag?: string | null;
  isShared?: boolean;
}

export interface UpdateJournalEntryRequest {
  isShared?: boolean;
}

export interface JournalEntriesResponse {
  entries: JournalEntry[];
  total: number;
}

export interface JournalPromptResponse {
  prompt: string;
}

export type GetMoodEntriesParams = {
limit?: number;
offset?: number;
};

export type GetJournalPromptParams = {
mood?: string;
};