# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  createdAt: Date;
}